refactor(mlt): move any logic out of index.js
